+++ solidity_tests/mapping2.sol
@@ -1,11 +1,10 @@
 contract Mapping {
     struct Participant {
-        uint total_paid;
-        uint payout;
+        uint256 total_paid;
+        uint256 payout;
     }
 
     mapping(address => Participant) participants;
-
 
     function deposit(address beneficiary) public payable {
         Participant storage b = participants[beneficiary];
@@ -13,7 +12,7 @@ contract Mapping {
         b.payout += msg.value / 2;
     }
 
-    function remove(uint amount, address beneficiary) public {
+    function remove(uint256 amount, address beneficiary) public {
         Participant storage b = participants[beneficiary];
         require(b.payout > 0);
         require(b.total_paid > 0);
@@ -26,5 +25,4 @@ contract Mapping {
         msg.sender.transfer(p.payout);
         p.payout = 0;
     }
-
 }
