+++ solidity_tests/recursive_mapping.sol
@@ -1,7 +1,7 @@
 contract RecursiveMapping {
     struct Participant {
-        uint total_paid;
-        mapping(address => uint) balances;
+        uint256 total_paid;
+        mapping(address => uint256) balances;
     }
 
     mapping(address => Participant) participants;
@@ -12,17 +12,20 @@ contract RecursiveMapping {
         p.balances[deposit_to] += msg.value;
     }
 
-    function transfer(uint amount, address take_from, address deposit_to) public {
+    function transfer(
+        uint256 amount,
+        address take_from,
+        address deposit_to
+    ) public {
         Participant storage p = participants[msg.sender];
         p.balances[take_from] -= amount;
         p.balances[deposit_to] += amount;
     }
 
-    function withdraw(address withdraw_from) public returns (int) {
+    function withdraw(address withdraw_from) public returns (int256) {
         Participant storage p = participants[withdraw_from];
         msg.sender.transfer(p.balances[msg.sender]);
         p.balances[msg.sender] = 0;
         return 42;
     }
-
 }
